test: Allow integration tests to run against a specific storageClass
In order to test changes to csi-digitalocean using DOKS clusters, we
need to deploy a second storage class that uses an updated version of
the CSI driver alongside the built-in version. Add an option in the
integration tests to specify a storage class, but still default to
`do-block-storage`.

While we're here, add an option to skip cleanup, which is helpful in
debugging failures.